requirements-local.txt
streamlit>=1.28.0
langgraph>=0.0.40
openai>=1.0.0
openpyxl>=3.1.0
pandas>=2.0.0
pillow>=10.0.0
pypdf2>=3.0.0
pyyaml>=6.0.0
trafilatura>=1.6.0
xlrd>=2.0.0